Add xl_auth.collection sub-package
