+++ xl_auth/collection/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""The collection module."""
+
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from . import views  # noqa
